feat: remove everything related to berlin
