--- src/data/apply/berlinFAQ.ts
@@ -1,59 +1,0 @@
-// Berlin-specific FAQ data
-export const berlinFaqs = [
-  {
-    question: "Do I need to be an AI expert to join TUM.ai?",
-    answer:
-      "You don't need to be an AI expert to join TUM.ai. We believe that individuals can learn and adapt quickly with the right mindset. Even if you're not an AI specialist but have a genuine interest in the field and are eager to learn and contribute, you could be a perfect fit for our team. After all, diverse perspectives and backgrounds can often bring fresh insights and innovative ideas. Note that the ML projects in your first semester are based on your area and level of expertise.",
-  },
-  {
-    question: "How does the application timeline look like?",
-    answer:
-      "Application March 24st - April 20th: Fill out our application form. Interview April 24th - April 26th: If you passed the screening phase, we will invite you to an interview round to get to know you better. Onboarding (Mandatory) May 16th - May 18th: Congratulations! You are invited to our Onboarding program to get to know TUM.ai!",
-  },
-  {
-    question: "How should I prepare myself for the interview?",
-    answer:
-      "To prepare for the interview, one of the fundamental steps is ensuring you have a comprehensive understanding of what TUM.ai stands for and its objectives. A deep dive into the TUM.ai website will offer insights into our vision, mission, projects, and the overall direction we're headed. By familiarizing yourself with the content on the webpage, not only will you be able to answer questions confidently, but you'll also be in a position to ask informed questions. Additionally, it might be helpful to keep up-to-date with the latest trends and developments in AI, as this could come up during your conversation. Remember, the more prepared you are, the more comfortable and confident you'll feel during the interview.",
-  },
-  {
-    question: "Do I need to relocate to Munich to join TUM.ai Berlin?",
-    answer:
-      "No! TUM.ai Berlin is an independent hub, and all activities, projects, and networking opportunities will take place in Berlin. However, Berlin members will be invited to Munich for the Onboarding Weekend in May and the Onboarding Getaway in June to connect with the broader TUM.ai community.",
-  },
-  {
-    question: "How will Berlin members collaborate with Munich members?",
-    answer:
-      "As a TUM.ai Berlin member, you will work on AI projects and task forces alongside members in Munich. While Munich members will meet in person, Berlin members will join remotely, just like our members currently abroad. You will not only benefit from our global network but also play a key role in shaping the Berlin hub.",
-  },
-  {
-    question: "Why is TUM.ai expanding to Berlin?",
-    answer:
-      "Berlin is one of Europe's leading tech and AI hubs, home to top research institutions, startups, and industry leaders. By expanding to Berlin, we are bridging the gap between two of Germany's most innovative ecosystems, Munich and Berlin, creating new opportunities for collaboration, knowledge exchange, and cutting-edge AI applications.",
-  },
-  {
-    question:
-      "Will the application process for Berlin be different from Munich?",
-    answer:
-      "No, the application process will follow the same selection criteria to ensure we bring in ambitious, like-minded individuals who align with TUM.ai's mission and values. The application timeline, interview format, and onboarding structure will remain consistent with Munich's process.",
-  },
-  {
-    question: "What kind of projects will be available in Berlin?",
-    answer:
-      "Berlin members will have access to the same diverse range of AI projects and research opportunities as Munich members. Whether you're interested in applied AI, deep tech research, entrepreneurship, or industry collaborations, you'll find projects that challenge you and push AI innovation forward.",
-  },
-  {
-    question: "How often will there be in-person events in Berlin?",
-    answer:
-      "TUM.ai Berlin will host regular meetups, AI talks, and networking events to build a strong, engaged community. Members will also have opportunities to travel to Munich for major TUM.ai events, hackathons, and summits.",
-  },
-  {
-    question: "Where can I contact TUM.ai Berlin for recruitment purposes?",
-    answer:
-      "If you have further questions, feel free to write us an email at berlin-contact@tum-ai.com.",
-  },
-  {
-    question: "Disclaimer",
-    answer:
-      "We are not an educational program. That means you are not only joining to learn but also to contribute to the development of the organization. We don't give in-depth lectures or crash courses in AI and coding yet - you need to interact with other members to learn things. Help is always given to those who ask. We do a lot of organizational work - the AI ecosystem is not ready, and you will be part of building it up. We have a membership fee of 10€ per semester for all of our active members.",
-  },
-];
